requirements.txt
fastapi
uvicorn
requests
PyMuPDF
python-multipart
google-generativeai
python-dotenv